+++ packages/ui/src/components/templates/report/index.ts
@@ -1,3 +1,2 @@
-export * from './adapters';
 export * from './components';
 export * from './hooks';