+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/workflow-core
 
+## 0.6.91
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.71
+
 ## 0.6.90
 
 ### Patch Changes