+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.91
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.91
+
 ## 0.6.90
 
 ### Patch Changes